requirements.txt
aiosqlite==0.20.0
aiohttp==3.9.5
discord.py==2.4.0
python-dotenv==1.0.1
